refactor(cli): make the split-type-identity check a shared validate check
It landed in fabric validate only, inside a section headed "everything
true of any pikku project" — which is where fabric validate keeps the
checks it inherits, not where shared checks are defined. Plain
pikku validate never ran it.

Nothing about the mechanism is fabric-specific. A dependency linked from
a sibling checkout resolves its own imports from that tree, and TS
follows symlinks to realpath — which is the normal setup for anyone
developing a pikku package against a consuming app, precisely the
population that hits this.

Now a check in the registry, so both entry points run it. Its precondition
is the root of an installed tree: a linked dependency is a property of the
install as a whole, so running per workspace package would report the same
pair once per package, and with no node_modules there is nothing to
compare.

930 tests pass.
